add bin2txt file
this commit adds a simple script to convert bin files to txt. CPU and GPU functions use different functions to do the same calculation.
this functions showed different precision, which causes solution and computed results to differ a little.
the added script is for debugging purposes and aims to help if the computed results are really wrong or just present a small deviation